Undo contract refactoring
