+++ starknet_py/tests/e2e/docs/guide/test_simple_deploy.py
@@ -7,7 +7,7 @@ from starknet_py.common import create_compiled_contract
 async def test_simple_deploy(account, map_class_hash, map_compiled_contract):
     # pylint: disable=import-outside-toplevel
     # docs: start
-    from starknet_py.net.contract.contract import Contract
+    from starknet_py.contract import Contract
 
     # docs: end
