+++ starknet_py/tests/e2e/docs/quickstart/test_using_contract.py
@@ -10,7 +10,7 @@ directory = os.path.dirname(__file__)
 async def test_using_contract(account, map_contract):
     # pylint: disable=unused-variable,too-many-locals
     # docs: start
-    from starknet_py.net.contract.contract import Contract
+    from starknet_py.contract import Contract
 
     contract_address = (
         "0x01336fa7c870a7403aced14dda865b75f29113230ed84e3a661f7af70fe83e7b"
