+++ starknet_py/tests/e2e/fixtures/accounts.py
@@ -6,11 +6,11 @@ from typing import Optional, Tuple
 import pytest
 import pytest_asyncio
 
+from starknet_py.contract import Contract
 from starknet_py.hash.address import compute_address
 from starknet_py.net.account.account import Account
 from starknet_py.net.account.base_account import BaseAccount
 from starknet_py.net.client_models import PriceUnit
-from starknet_py.net.contract.contract import Contract
 from starknet_py.net.full_node_client import FullNodeClient
 from starknet_py.net.http_client import HttpMethod, RpcHttpClient
 from starknet_py.net.models import StarknetChainId
