+++ starknet_py/tests/e2e/proxy/proxy_test.py
@@ -2,10 +2,10 @@ from typing import Optional
 
 import pytest
 
+from starknet_py.contract import Contract
 from starknet_py.hash.storage import get_storage_var_address
 from starknet_py.net.client import Client
 from starknet_py.net.client_errors import ContractNotFoundError
-from starknet_py.net.contract.contract import Contract
 from starknet_py.net.models import Address
 from starknet_py.proxy.contract_abi_resolver import ProxyResolutionError
 from starknet_py.proxy.proxy_check import ProxyCheck
